+++ __tests__/auth.test.ts
@@ -2,6 +2,7 @@ import { SuperAgentTest } from "supertest";
 import { Player, PrismaClient } from "@prisma/client";
 import { BAD_REQUEST, FORBIDDEN, UNAUTHORIZED } from "http-status";
 import jwt from "jsonwebtoken";
+// import * as nodemailer from "nodemailer";
 import { initAgent } from "./helpers";
 import { TokenPair } from "@/types/response/jwt";
 import { AuthServices } from "@/components/auth/services";
@@ -21,6 +22,10 @@ const credentials = {
   username: "jest_test" + Date.now(),
   password: "1234",
   email: "jest_test" + Date.now() + "@test.com",
+};
+const passwordResetRequest = {
+  new_password: "12345",
+  repeat_password: "12345",
 };
 
 beforeAll(initialize);
@@ -135,6 +140,110 @@ describe("[UNIT] => AUTH", () => {
       expect(response.status).toBe(FORBIDDEN);
     });
   });
+
+  describe("POST: /auth/reset-password", () => {
+    it("Should return 404", async () => {
+      const response = await agent
+        .post(`/app/${CONFIG.APP.VER}/auth/reset-password`)
+        .send(passwordResetRequest)
+        .set("Authorization", `Bearer ${tokenPair2.access}`)
+        .set("User-Agent", USER_AGENT);
+
+      expect(response.status).toBe(404);
+    });
+
+    it.each`
+      new_password      | repeat_password   | message
+      ${"12345"}        | ${"123456"}       | ${"repeat_password must match new_password"}
+      ${"12"}           | ${"12"}           | ${"password must be at least 4 characters long"}
+      ${"a".repeat(73)} | ${"a".repeat(73)} | ${"password must be under 73 characters"}
+      ${""}             | ${"12345"}        | ${"new_password is required"}
+      ${"12345"}        | ${""}             | ${"repeat_password is required"}
+    `(
+      "Should return 400",
+      async ({ new_password, repeat_password, message }) => {
+        const response = await agent
+          .post(`/app/${CONFIG.APP.VER}/auth/reset-password`)
+          .send({
+            new_password,
+            repeat_password,
+          })
+          .set("Authorization", `Bearer ${tokenPair2.access}`)
+          .set("User-Agent", USER_AGENT);
+
+        expect(response.status).toBe(BAD_REQUEST);
+        expect(response.body.data[0].msg).toBe(message);
+      },
+    );
+
+    it("Should return 401", async () => {
+      const response = await agent
+        .post(`/app/${CONFIG.APP.VER}/auth/reset-password`)
+        .send(passwordResetRequest);
+
+      expect(response.status).toBe(UNAUTHORIZED);
+    });
+  });
+
+  describe("POST: /auth/forgot-password", () => {
+    // const sendMailMock = jest.fn();
+    // jest.mock("nodemailer", () => ({
+    //   createTransport: jest.fn().mockImplementation(() => ({
+    //     sendMail: sendMailMock,
+    //   })),
+    // }));
+
+    // it.only("Should send password reset email", async () => {
+    //   const newPasswordServices = new NewPasswordServices();
+    //   await newPasswordServices.forgotPassword(player.username);
+
+    //   expect(nodemailer.createTransport().sendMail).toHaveBeenCalledTimes(1);
+    // });
+
+    it("Should return 200 with valid username", async () => {
+      const response = await agent
+        .post(`/app/${CONFIG.APP.VER}/auth/forgot-password`)
+        .send({ username: player.username });
+
+      expect(response.status).toBe(200);
+    });
+
+    it("Should return 200 with invalid username", async () => {
+      const response = await agent
+        .post(`/app/${CONFIG.APP.VER}/auth/forgot-password`)
+        .send({ username: "non_existing_username" });
+
+      expect(response.status).toBe(200);
+    });
+
+    it("Should return 400 unknown_fields", async () => {
+      const response = await agent
+        .post(`/app/${CONFIG.APP.VER}/auth/forgot-password`)
+        .send({ unknown_field: "" })
+        .send({ username: "batata" });
+
+      expect(response.status).toBe(400);
+      expect(response.body.data[0].type).toBe("unknown_fields");
+    });
+
+    it("Should return 400 username_required", async () => {
+      const response = await agent.post(
+        `/app/${CONFIG.APP.VER}/auth/forgot-password`,
+      );
+
+      expect(response.status).toBe(400);
+      expect(response.body.data[0].msg).toBe("username is required");
+    });
+
+    it("Should return 429", async () => {
+      const response = await agent
+        .post(`/app/${CONFIG.APP.VER}/auth/forgot-password`)
+        .send({ username: player.username });
+
+      expect(response.status).toBe(429);
+      expect(response.body.code).toBe("too_many_requests");
+    });
+  });
 });
 
 async function initialize() {
@@ -153,6 +262,9 @@ async function initialize() {
       email: "jest_test2" + Date.now() + "@test.com",
       password: "1234",
       panel_id: -20,
+      roles: {
+        connect: { name: CONFIG.ROLES.PLAYER },
+      },
     },
   });